moved svelte component to react
